Update functional_specification.md
Add all details on the mentor system from the design and it's different functionalities it has.

Add Informations about the saving system, it includes when will the game save automatically and how to access those saves.

delete "6. Objectives & Freedom"  as it became irrelevant since the different objectives and allowed freedom are mentioned in Core Game Loops, Mentor System and Game Repairs sections.

Add a description on the different Audios and Sound Design.